feat: move serial port into new file
